# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Video Topic Graph Platform API
  description: |
    Open-source platform for video topic graph generation.
    All endpoints require OIDC authentication via Bearer token.
  version: 1.0.0
  contact:
    name: Video Topic Graph Platform
    url: https://github.com/video-topic-graph/platform
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /exports:
    post:
      summary: Create an export
      description: Generate PPTX, HTML, or PDF export
      tags: [Exports]
      operationId: createExport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExportRequest'
      responses:
        '202':
          description: Export job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
  /exports/{id}:
    get:
      summary: Get export status and download URL
      tags: [Exports]
      operationId: getExport
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Export details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /exports/{id}/download:
    get:
      summary: Download export file
      tags: [Exports]
      operationId: downloadExport
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Export file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC Bearer token from Keycloak
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    CreateExportRequest:
      type: object
      required: [video_id, type]
      properties:
        video_id:
          type: string
          format: uuid
        graph_version_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [pptx, html, pdf]
        options:
          type: object
          properties:
            include_snippets:
              type: boolean
              default: true
            snippet_embed_mode:
              type: string
              enum: [embedded, linked, none]
              default: linked
            template:
              type: string
              enum: [default, minimal, detailed]
              default: default
            topic_levels:
              type: array
              items:
                type: integer
            include_appendix:
              type: boolean
              default: true
...
    ExportResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        video_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [pptx, html, pdf]
        status:
          type: string
          enum: [pending, processing, complete, error]
        storage_path:
          type: string
          nullable: true
        download_url:
          type: string
          format: uri
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
          nullable: true
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
        request_id:
          type: string